Pull evaluation_parameter_store out into its own place
