+++ great_expectations/data_context/store/__init__.py
@@ -9,11 +9,14 @@ from .store import (
     WriteOnlyStore,
     ReadWriteStore,
     BasicInMemoryStore,
-    EvaluationParameterStore,
 )
 
 from .namespaced_read_write_store import (
     NamespacedReadWriteStore,
     ValidationResultStore,
     ExpectationStore,
+)
+
+from .evaluation_parameter_store import (
+    EvaluationParameterStore,
 )
