+++ great_expectations/data_context/store/store.py
@@ -264,21 +264,3 @@ class BasicInMemoryStore(ReadWriteStore):
 #                 key,
 #                 type(key),
 #             ))
-
-class EvaluationParameterStore(object):
-    """You want to be a dict. You get to be a dict. But we call you a Store."""
-    
-    def __init__(self, root_directory=None):
-        self.store = {}
-
-    def get(self, key):
-        return self.store[key]
-
-    def set(self, key, value):
-        self.store[key] = value
-
-    def has_key(self, key):
-        return key in self.store
-
-    def list_keys(self):
-        return list(self.store.keys())
